requirements.txt
spacy
spacy-model-en_core_web_sm
json
tqdm
pytorch
itertools
collections
gensim
pickle
flask
